pyproject.toml
[project]
name = "sip-pbx"
version = "0.1.0"
description = "SIP PBX with Real-time Voice Analysis"
authors = [
    {name = "SIP PBX Team"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["sip", "pbx", "b2bua", "voice-analysis", "ai", "stt", "emotion-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # SIP Stack
    "pjsua2>=2.13",
    
    # AI/ML
    "faster-whisper>=1.0.0",
    "torch>=2.1.0",
    "onnxruntime-gpu>=1.16.0",
    "transformers>=4.35.0",
    "speechbrain>=1.0.0",
    "librosa>=0.10.0",
    
    # Async & Web
    "aiohttp>=3.9.1",
    "asyncio>=3.4.3",
    
    # Configuration & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
    
    # Audio Processing
    "av>=11.0.0",
    
    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "structlog>=24.0.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "types-PyYAML",
    "pip-audit>=2.6.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = [
    "tests",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "benchmark: Performance benchmark tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

